requirements.txt
streamlit
pandas
jinja2
json
xhtml2pdf
stqdm